refactor: separate controller
